+++ src/View/Helper/LibraryHelper.php
@@ -88,19 +88,6 @@ class LibraryHelper extends Helper
     }
 
     /**
-     * Create a script block for Google Analytics
-     * @param string $id Analytics ID
-     * @return string|null A script tag or `null`
-     * @deprecated 2.23.1 It will be removed in a later release
-     */
-    public function analytics(string $id): ?string
-    {
-        deprecationWarning('Deprecated. It will be removed in a later release');
-
-        return $this->getView()->getRequest()->is('localhost') ? null : $this->Html->scriptBlock('!function(e,a,t,n,c,o,s){e.GoogleAnalyticsObject=c,e[c]=e[c]||function(){(e[c].q=e[c].q||[]).push(arguments)},e[c].l=1*new Date,o=a.createElement(t),s=a.getElementsByTagName(t)[0],o.async=1,o.src=n,s.parentNode.insertBefore(o,s)}(window,document,"script","//www.google-analytics.com/analytics.js","ga"),ga("create","' . $id . '","auto"),ga("send","pageview");', ['block' => 'script_bottom']);
-    }
-
-    /**
      * Loads all CKEditor scripts.
      *
      * To know how to install and configure CKEditor, please refer to the `README.md` file.